Session(ScreenMirroring) Commit#20

+++ src/sample/Main.java
@@ -1,6 +1,5 @@
 package sample;
 
-import Utils.JGradientButton;
 import javafx.application.Application;
 import javafx.application.Platform;
 import javafx.fxml.FXMLLoader;
@@ -9,59 +8,8 @@ import javafx.scene.Scene;
 import javafx.scene.image.Image;
 import javafx.stage.Stage;
 
-import javax.swing.*;
-import java.awt.*;
-import java.awt.event.MouseEvent;
-import java.awt.event.MouseListener;
-import java.awt.event.MouseMotionListener;
 import java.io.File;
 import java.util.ArrayList;
-
-class Listener implements MouseMotionListener, MouseListener{
-
-    int x = -1;
-    int y = -1;
-
-    JFrame j;
-
-    Listener(JFrame j){
-        this.j = j;
-    }
-
-    @Override
-    public void mouseClicked(MouseEvent e) {}
-
-    @Override
-    public void mousePressed(MouseEvent e) {
-        x = e.getPoint().x;
-        y = e.getPoint().y;
-    }
-
-    @Override
-    public void mouseReleased(MouseEvent e) {
-    }
-
-    @Override
-    public void mouseEntered(MouseEvent e) {
-
-    }
-
-    @Override
-    public void mouseExited(MouseEvent e) {
-
-    }
-
-    @Override
-    public void mouseDragged(MouseEvent e) {
-        j.setLocation(j.getLocation().x - x + e.getPoint().x, j.getLocation().y - y + e.getPoint().y);
-
-    }
-
-    @Override
-    public void mouseMoved(MouseEvent e) {
-
-    }
-}
 
 public class Main extends Application {
 
@@ -82,32 +30,14 @@ public class Main extends Application {
 
     @Override
     public void stop() throws Exception{
-        for(File f : tempfiles) {
-            f.deleteOnExit();
-        }
+        if(tempfiles != null)
+            for(File f : tempfiles) {
+                f.deleteOnExit();
+            }
     }
 
 
     public static void main(String[] args) throws InterruptedException {
-        JFrame f = new JFrame("MouseListener");
-        f.setSize(50,65);
-        f.setAlwaysOnTop(true);
-        f.setUndecorated(true);
-        f.setVisible(true);
-        JPanel p = new JPanel();
-        Listener l = new Listener(f);
-        f.addMouseMotionListener(l);
-        f.addMouseListener(l);
-        JGradientButton SwapButton = new JGradientButton("Swap");
-        JButton FinishButton = new JButton("Finish");
-        p.add(SwapButton);
-        p.add(FinishButton);
-        p.setLayout(new FlowLayout());
-        f.setBackground(Color.getColor("#202020"));
-        f.add(p);
-        f.setDefaultCloseOperation(JFrame.DO_NOTHING_ON_CLOSE);
-
-        f.show();
         launch(args);
     }
 }
